remember commit

+++ add_with_1_function.c
@@ -2,12 +2,13 @@
 #include<stdio.h>
 int main()
 {
-	int a, b, sum;
+	float a, b, sum;
 	printf("Input\nEnter number: \n");	
-	scanf("%d",&a);
+	scanf("%f",&a);
 	printf("Enter number: \n");
-	scanf("%d",&b);
+	scanf("%f",&b);
 	sum=a+b;
-	printf("Output\nSum of %d and %d is %d.\n",a,b,sum);
+	printf("Output\nSum of %f and %f is %f.\n",a,b,sum);
 	return 0;
 }
+
